requirements.txt
aiohttp>=3.9
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
statsmodels>=0.14
python-dotenv>=1.0